+++ wallets/metamask/src/fixture-actions/prepareExtension.ts
@@ -1,4 +1,4 @@
-import { downloadFile, ensureCacheDirExists, unzipArchive } from '@synthetixio/synpress-core'
+import { downloadFile, ensureCacheDirExists, unzipArchive } from '../../../../packages/cache'
 
 export const DEFAULT_METAMASK_VERSION = '11.9.1'
 export const EXTENSION_DOWNLOAD_URL = `https://github.com/MetaMask/metamask-extension/releases/download/v${DEFAULT_METAMASK_VERSION}/metamask-chrome-${DEFAULT_METAMASK_VERSION}.zip`