+++ wallets/metamask/src/fixtures/metaMaskFixtures.ts
@@ -8,7 +8,7 @@ import {
   createTempContextDir,
   defineWalletSetup,
   removeTempContextDir
-} from '@synthetixio/synpress-core'
+} from '../../../../packages/cache'
 import { type Anvil, type CreateAnvilOptions, createPool } from '@viem/anvil'
 import fs from 'fs-extra'