+++ wallets/metamask/test/wallet-setup/basic.setup.ts
@@ -1,4 +1,4 @@
-import { defineWalletSetup } from '@synthetixio/synpress-core'
+import { defineWalletSetup } from '../../../../packages/cache'
 import { MetaMask } from '../../src'
 
 const SEED_PHRASE = 'test test test test test test test test test test test junk'