+++ release/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Alpha Release
 - Updated dependencies
-  - @synthetixio/synpress-core@0.0.1-alpha.3
+  - @synthetixio/synpress-cache@0.0.1-alpha.3
   - @synthetixio/synpress-test@0.0.1-alpha.3
   - @synthetixio/synpress-metamask@0.0.1-alpha.3
 
@@ -16,7 +16,7 @@
 
 - Alpha Release
 - Updated dependencies
-  - @synthetixio/synpress-core@0.0.1-alpha.2
+  - @synthetixio/synpress-cache@0.0.1-alpha.2
   - @synthetixio/synpress-test@0.0.1-alpha.2
   - @synthetixio/synpress-metamask@0.0.1-alpha.2
 
@@ -26,7 +26,7 @@
 
 - Alpha Release
 - Updated dependencies
-  - @synthetixio/synpress-core@0.0.1-alpha.1
+  - @synthetixio/synpress-cache@0.0.1-alpha.1
   - @synthetixio/synpress-test@0.0.1-alpha.1
   - @synthetixio/synpress-metamask@0.0.1-alpha.1